third_party/WebKit/LayoutTests/inspector/sources/debugger/debugger-pause-on-blocked-script-injection-expected.txt

CONSOLE ERROR: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self'". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.

Tests that pause on exception works for inline scripts blocked by CSP.

Debugger was enabled.
Script execution paused.
Call stack:
    0) testAction (pause-on-blocked-script-injection.js:5)
Script execution resumed.
Debugger was disabled.
